fillers

+++ src/pizzapazza/filler/Z4ConicFiller.java
@@ -11,8 +11,8 @@ import simulation.js.$Object;
  */
 public class Z4ConicFiller extends Z4AbstractFiller {
 
-  private final double cx;
-  private final double cy;
+  private final int cx;
+  private final int cy;
   private final double angle;
   private final boolean symmetric;
 
@@ -20,14 +20,12 @@ public class Z4ConicFiller extends Z4AbstractFiller {
    * Creates the object
    *
    * @param gradientColor The color used to fill
-   * @param cx The x-axis coordinate of the center point in relative size (in
-   * the range [0,1])
-   * @param cy The y-axis coordinate of the center point in relative size (in
-   * the range [0,1])
+   * @param cx The x-axis coordinate of the center point
+   * @param cy The y-axis coordinate of the center point
    * @param angle The rotation angle of the cone (in radians)
    * @param symmetric true for symmetric cone, false otherwise
    */
-  public Z4ConicFiller(Z4GradientColor gradientColor, double cx, double cy, double angle, boolean symmetric) {
+  public Z4ConicFiller(Z4GradientColor gradientColor, int cx, int cy, double angle, boolean symmetric) {
     super(gradientColor);
 
     this.cx = cx;
@@ -37,7 +35,7 @@ public class Z4ConicFiller extends Z4AbstractFiller {
   }
 
   @Override
-  protected double getColorPositionAt(double x, double y) {
+  protected double getColorPositionAt(int x, int y) {
     $Object rotated = Z4Math.rotate(x - this.cx, y - this.cy, this.angle);
     double position = Math.atan2((double) rotated.$get("y"), (double) rotated.$get("x")) / Z4Math.TWO_PI;
